Elastic_search/requirements.txt

elasticsearch==9.0.1
pandas==2.2.3
python-dotenv==1.1.0
